cognitive_llm/evaluation/ablation.py
"""
Ablation study runner — systematically enables/disables blocks to measure contribution.

Implements the ablation matrix from the spec to isolate each block's contribution.
"""

import itertools
from dataclasses import dataclass, field
from pathlib import Path

import yaml

try:
    import wandb
except ImportError:
    wandb = None


@dataclass
class AblationExperiment:
    """Defines a single ablation experiment."""

    name: str
    use_block1: bool = False
    use_block2: bool = False
    use_block3: bool = False
    use_block4: bool = False  # Phase 2 only
    use_block5: bool = False
    use_block6: bool = True  # Always ON when any block is ON

    def to_config(self) -> dict:
        return {
            "use_block1": self.use_block1,
            "use_block2": self.use_block2,
            "use_block3": self.use_block3,
            "use_block4": self.use_block4,
            "use_block5": self.use_block5,
            "use_block6": self.use_block6,
        }

    def block_string(self) -> str:
        """Return a string like '100001' representing active blocks."""
        return "".join(
            str(int(getattr(self, f"use_block{i}")))
            for i in range(1, 7)
        )


# Standard ablation matrix from spec
ABLATION_MATRIX: list[AblationExperiment] = [
    AblationExperiment(
        name="baseline",
        use_block6=False,
    ),
    AblationExperiment(
        name="+B1_only",
        use_block1=True,
        use_block6=True,
    ),
    AblationExperiment(
        name="+B2_only",
        use_block2=True,
        use_block6=True,
    ),
    AblationExperiment(
        name="+B3_only",
        use_block3=True,
        use_block6=True,
    ),
    AblationExperiment(
        name="+B1+B2",
        use_block1=True,
        use_block2=True,
        use_block6=True,
    ),
    AblationExperiment(
        name="full_B1-3_5_6",
        use_block1=True,
        use_block2=True,
        use_block3=True,
        use_block5=True,
        use_block6=True,
    ),
]


@dataclass
class AblationResult:
    """Stores results of one ablation experiment."""

    experiment: AblationExperiment
    metrics: dict = field(default_factory=dict)
    train_losses: list[float] = field(default_factory=list)
    checkpoint_path: str | None = None


class AblationRunner:
    """
    Runs the full ablation study across all experiment configurations.

    Args:
        base_config: Base training config (YAML dict).
        model_loader: Callable that returns a fresh base model.
        trainer_cls: Trainer class to use for training.
        benchmark_runner: BenchmarkRunner instance for final scoring.
        output_dir: Directory to save ablation results.
    """

    def __init__(
        self,
        base_config: dict,
        model_loader,
        trainer_cls,
        benchmark_runner=None,
        output_dir: str = "./ablation_results",
    ):
        self.base_config = base_config
        self.model_loader = model_loader
        self.trainer_cls = trainer_cls
        self.benchmark_runner = benchmark_runner
        self.output_dir = Path(output_dir)
        self.output_dir.mkdir(parents=True, exist_ok=True)
        self.results: list[AblationResult] = []

    def run_all(
        self, experiments: list[AblationExperiment] | None = None
    ) -> list[AblationResult]:
        """
        Run all ablation experiments.

        Args:
            experiments: List of experiments. Defaults to ABLATION_MATRIX.

        Returns:
            List of AblationResult objects.
        """
        if experiments is None:
            experiments = ABLATION_MATRIX

        for exp in experiments:
            print(f"\n{'='*60}")
            print(f"Ablation Experiment: {exp.name}")
            print(f"Blocks: {exp.block_string()}")
            print(f"{'='*60}")

            result = self.run_experiment(exp)
            self.results.append(result)

        self._save_summary()
        return self.results

    def run_experiment(self, experiment: AblationExperiment) -> AblationResult:
        """
        Run a single ablation experiment.

        Args:
            experiment: AblationExperiment configuration.

        Returns:
            AblationResult with metrics and losses.
        """
        # Merge block config with base training config
        config = {**self.base_config, **experiment.to_config()}

        # Load fresh model and create cognitive wrapper
        from cognitive_llm.models.cognitive_model import CognitiveModel

        base_model = self.model_loader()
        model = CognitiveModel(base_model, config)

        # Create trainer and train
        # Note: actual dataloader creation is left to the caller
        # This is a simplified interface
        result = AblationResult(experiment=experiment)

        if self.benchmark_runner:
            # Run benchmarks on the trained model
            benchmark_results = self.benchmark_runner.run_all()
            for br in benchmark_results:
                result.metrics[f"{br.task}/{br.metric}"] = br.score

        return result

    def _save_summary(self) -> None:
        """Save ablation summary to YAML file."""
        summary = {}
        for result in self.results:
            summary[result.experiment.name] = {
                "blocks": result.experiment.block_string(),
                "metrics": result.metrics,
            }

        summary_path = self.output_dir / "ablation_summary.yaml"
        with open(summary_path, "w") as f:
            yaml.dump(summary, f, default_flow_style=False)

        print(f"\nAblation summary saved to {summary_path}")

    def print_comparison(self) -> None:
        """Print a comparison table of all ablation results."""
        if not self.results:
            print("No results to compare.")
            return

        # Collect all unique metrics
        all_metrics = set()
        for r in self.results:
            all_metrics.update(r.metrics.keys())

        metrics_list = sorted(all_metrics)

        # Header
        header = f"{'Experiment':<20} {'Blocks':<8}"
        for m in metrics_list:
            header += f" {m:<15}"
        print("\n" + "=" * len(header))
        print(header)
        print("=" * len(header))

        # Rows
        for r in self.results:
            row = f"{r.experiment.name:<20} {r.experiment.block_string():<8}"
            for m in metrics_list:
                score = r.metrics.get(m, float("nan"))
                row += f" {score:<15.4f}"
            print(row)

        print("=" * len(header))
