requirements.txt
torch>=2.3.0
transformers>=4.53.0
datasets>=2.18.0
accelerate>=0.30.0
bitsandbytes>=0.43.0
peft>=0.10.0
trl>=0.8.6
evaluate>=0.4.0
scipy>=1.11.0
numpy>=1.26.0
wandb>=0.16.0
lm-eval>=0.4.2
pyyaml>=6.0
